integrate VMD and SYVA
VMD is molecular graphics software used for visualizing and analyzing molecular structures. It enables the simultaneous display of multiple structures with various rendering styles and color schemes. The Symmetry Tool plugin within VMD is particularly useful for analyzing point group symmetry in small molecules, providing detailed information on principal axes, reflections, and symmetry planes. While this plugin utilizes the compiled code, it may encounter challenges in determining point groups for certain molecules with T, Th, or O symmetry. On the other hand, SYVA is a program that employs vector algebra to efficiently analyze molecular symmetry efficiently, with a high rate of accuracy in identifying point groups. To enhance the display of molecular symmetry, we have made modifications to both the Symmetry Tool plugin and the SYVA program.